compiler: Move and enhance FunctionMap
